deepmind-veo/requirements.txt
python-dotenv
google-genai==1.30.0
